requirements.txt
django
djangorestframework
Pillow
djangorestframework-simplejwt
